fix format of numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,18 +7,18 @@
  */
 int main(void)
 {
-	int next;
-	int first = 1;
-	int second = 2;
-	int count;
+	long int next;
+	long int first = 1;
+	long int second = 2;
+	long int count;
 
-	printf("%d, %d, ", first, second);
+	printf("%ld, %ld, ", first, second);
 	for (count = 2; count < 50; count++)
 	{
 		next = first + second;
 		first = second;
 		second = next;
-		printf("%d, ", next);
+		printf("%ld, ", next);
 	}
 	printf("\n");
 	return (0);
